app/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
djoser==2.2.0
python-dateutil==2.8.2
celery==5.2.7
flower==1.0.0
gunicorn==20.1.0
psycopg2==2.9.3
redis==5.0.1
python-dotenv==0.20.0
drf-yasg==1.21.4
django-rest-passwordreset==1.3.0
django-celery-beat==2.5.0
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
Brotli==1.0.9
certifi==2023.5.7
charset-normalizer==3.1.0
colorama==0.4.6
discum==1.4.1
filetype==1.2.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
openai==0.27.8
requests==2.31.0
requests-toolbelt==1.0.0
six==1.16.0
tqdm==4.65.0
ua-parser==0.16.1
urllib3==1.26.16
websocket-client==0.59.0
yarl==1.9.2
django_cors_headers==4.1.0
boto3==1.26.165
botocore==1.29.165
mixpanel==4.10.0
deep-translator
cyrtranslit
django-nested-admin
django-debug-toolbar
nltk
python-telegram-bot==13.13
stripe==6.0.0
docker==6.1.3
flower==1.0.0
django-ses==3.5.0
Pillow==10.1.0
